Do not depend on clock
This is trading correctness for fewer dependencies.  One immediate
benefit of this change is that installation for GHC 9.2.1 will succeed
without installing `libnuma-dev` first.

The new code uses getCurrentTime, which is not monotonic.  As a
consequence the new behavior is less accurate when:

- A test run spans a leap second
- The system clock is adjusted during a test run

Both of these events are reasonably rare, so I hope this is the right
thing to do.